rename slider to carousel
