+++ app/src/app/pages/artwork/artwork.component.spec.ts
@@ -1,8 +1,8 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { ArtworkComponent } from './artwork.component';
-import { SlideComponent } from 'src/app/shared/components/slider/slide/slide.component';
-import { SliderComponent } from 'src/app/shared/components/slider/slider.component';
+import { SlideComponent } from 'src/app/shared/components/carousel/slide/slide.component';
+import { CarouselComponent } from 'src/app/shared/components/carousel/carousel.component';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { DataService } from 'src/app/core/services/data.service';
 import { HttpClientModule } from '@angular/common/http';
@@ -24,7 +24,7 @@ describe('ArtworkComponent', () => {
       declarations: [
         ArtworkComponent,
         SlideComponent,
-        SliderComponent,
+        CarouselComponent,
       ],
       providers: [
         DataService,
