+++ app/src/app/shared/components/carousel/carousel.component.html
