+++ app/src/app/shared/components/carousel/carousel.component.spec.ts
@@ -1,6 +1,6 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
-import { SliderComponent } from './slider.component';
+import { CarouselComponent } from './carousel.component';
 import { SlideComponent } from './slide/slide.component';
 import { DataService } from 'src/app/core/services/data.service';
 import { RouterModule } from '@angular/router';
@@ -8,8 +8,8 @@ import { HttpClientModule } from '@angular/common/http';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 
 describe('SliderComponent', () => {
-  let component: SliderComponent;
-  let fixture: ComponentFixture<SliderComponent>;
+  let component: CarouselComponent;
+  let fixture: ComponentFixture<CarouselComponent>;
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
@@ -19,7 +19,7 @@ describe('SliderComponent', () => {
         RouterModule.forRoot([])
       ],
       declarations: [
-        SliderComponent,
+        CarouselComponent,
         SlideComponent,
       ],
       providers: [
@@ -29,7 +29,7 @@ describe('SliderComponent', () => {
   }));
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(SliderComponent);
+    fixture = TestBed.createComponent(CarouselComponent);
     component = fixture.componentInstance;
     fixture.detectChanges();
   });
