+++ app/src/app/shared/components/carousel/slide/slide.component.scss
