+++ app/src/app/shared/components/carousel/slide/slide.component.ts
@@ -1,7 +1,7 @@
 import { Component, Input, Output, EventEmitter, AfterViewInit, ElementRef } from '@angular/core';
 import { Entity } from 'src/app/shared/models/models';
 import { DataService } from 'src/app/core/services/data.service';
-import { Slide, makeDefaultSlide } from '../slider.component';
+import { Slide, makeDefaultSlide } from '../carousel.component';
 
 /**
  * a slide of the slider.
