+++ app/src/app/shared/components/search/search.component.spec.ts
@@ -1,8 +1,8 @@
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 
 import { SearchComponent } from './search.component';
-import { SlideComponent } from '../slider/slide/slide.component';
-import { SliderComponent } from '../slider/slider.component';
+import { SlideComponent } from '../carousel/slide/slide.component';
+import { CarouselComponent } from '../carousel/carousel.component';
 import { DataService } from 'src/app/core/services/data.service';
 import { RouterModule } from '@angular/router';
 import { HttpClientModule } from '@angular/common/http';
@@ -24,7 +24,7 @@ describe('SearchComponent', () => {
       declarations: [
         SearchComponent,
         SlideComponent,
-        SliderComponent,
+        CarouselComponent,
       ],
       providers: [
         DataService,
