+++ app/src/app/shared/shared.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { SliderComponent } from './components/slider/slider.component';
-import { SlideComponent } from './components/slider/slide/slide.component';
+import { CarouselComponent } from './components/carousel/carousel.component';
+import { SlideComponent } from './components/carousel/slide/slide.component';
 import { NgbModule } from '@ng-bootstrap/ng-bootstrap';
 import { RouterModule } from '@angular/router';
 import { SearchComponent } from './components/search/search.component';
@@ -9,8 +9,8 @@ import { FormsModule } from '@angular/forms';
 
 /** Everything that should be used within multiple feature modules but isn't always required goes here */
 @NgModule({
-  declarations: [SliderComponent, SlideComponent, SearchComponent],
+  declarations: [CarouselComponent, SlideComponent, SearchComponent],
   imports: [CommonModule, NgbModule, RouterModule, FormsModule],
-  exports: [SliderComponent, SearchComponent],
+  exports: [CarouselComponent, SearchComponent],
 })
 export class SharedModule {}
